+++ src/commands/owner.rs
@@ -25,7 +25,7 @@ pub async fn quit(ctx: Context<'_>) -> Result<(), Error> {
 /// Manage admins which are able to modify the server.
 #[poise::command(
     slash_command,
-    check = "crate::checks::is_owner",
+    check = "crate::checks::is_admin",
     subcommands("add", "remove", "list")
 )]
 pub async fn admin(_ctx: Context<'_>) -> Result<(), Error> {
@@ -34,7 +34,7 @@ pub async fn admin(_ctx: Context<'_>) -> Result<(), Error> {
 }
 
 /// Manage admins which are able to modify the server.
-#[poise::command(slash_command, check = "crate::checks::is_owner")]
+#[poise::command(slash_command, check = "crate::checks::is_admin")]
 pub async fn add(
     ctx: Context<'_>,
     #[description = "User to add to the admin list"] user: User,
@@ -49,7 +49,7 @@ pub async fn add(
 }
 
 /// Manage admins which are able to modify the server.
-#[poise::command(slash_command, check = "crate::checks::is_owner")]
+#[poise::command(slash_command, check = "crate::checks::is_admin")]
 pub async fn remove(
     ctx: Context<'_>,
     #[description = "User to remove from the admin list"] user: User,
@@ -64,7 +64,7 @@ pub async fn remove(
 }
 
 /// Manage admins which are able to modify the server.
-#[poise::command(slash_command, check = "crate::checks::is_owner")]
+#[poise::command(slash_command, check = "crate::checks::is_admin")]
 pub async fn list(ctx: Context<'_>) -> Result<(), Error> {
     let state = ctx.data().state.lock().await;
 
@@ -85,7 +85,12 @@ pub async fn list(ctx: Context<'_>) -> Result<(), Error> {
 /// Register application commands in this guild or globally
 ///
 /// Run with no arguments to register in guild, run with argument "global" to register globally.
-#[poise::command(prefix_command, hide_in_help, check = "crate::checks::is_owner")]
+#[poise::command(
+    prefix_command,
+    dm_only,
+    hide_in_help,
+    check = "crate::checks::is_owner"
+)]
 pub async fn register(ctx: Context<'_>) -> Result<(), Error> {
     poise::builtins::register_application_commands_buttons(ctx).await?;
